+++ src/api/baseUrl.ts
@@ -1,0 +1,3 @@
+const baseUrl: string = "http://localhost:8080";
+
+export { baseUrl };
