+++ src/services/constant.ts
@@ -1,2 +1,2 @@
-export const token =
-  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c2VyLTEyMyIsInVzZXJuYW1lIjoibWFuYWdlciIsInJvbGUiOiJtYW5hZ2VyIiwiaWF0IjoxNzQ2MjQ5NDEzfQ.75kGrkaOTUqu1ok64mNBKRJLOrxtLcQx4Oo33e_VQqM";
+// export const token =
+//   "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c2VyLTEyMyIsInVzZXJuYW1lIjoibWFuYWdlciIsInJvbGUiOiJtYW5hZ2VyIiwiaWF0IjoxNzQ2MjQ5NDEzfQ.75kGrkaOTUqu1ok64mNBKRJLOrxtLcQx4Oo33e_VQqM";
